+++ packages/core-p2p/src/utils/is-whitelisted.ts
@@ -2,14 +2,11 @@ import mm from "micromatch";
 
 /**
  * Check if the given IP address is whitelisted.
- * @param  {[]String} value
- * @param  {String} value
- * @return {boolean}
  */
-export = (whitelist, value) => {
+export const isWhitelisted = (whitelist: string[], ip: string): boolean => {
     if (Array.isArray(whitelist)) {
         for (const item of whitelist) {
-            if (mm.isMatch(value, item)) {
+            if (mm.isMatch(ip, item)) {
                 return true;
             }
         }
