+++ packages/core-p2p/src/utils/restore-peers.ts
@@ -1,0 +1,10 @@
+import { existsSync } from "fs";
+import { config } from "../config";
+
+export const restorePeers = (): void => {
+    const path = `${process.env.CORE_PATH_CACHE}/peers.json`;
+    if (existsSync(path)) {
+        const peers = require(path);
+        config.set("peers", peers);
+    }
+};
